changed alt from int16_t to int32_t changed alt in mavlink to cm

+++ open360tracker/telemetry.h
@@ -8,7 +8,7 @@
 
 
 //alt in meter
-int16_t getTargetAlt();
+int32_t getTargetAlt();
 void encodeTargetData(uint8_t c);
 
 #ifdef MFD
